# Speed up build metric name.